Move input error login into class InputWithError
